Handles additional config file
This patch adds the ability to add an additional config file.
The intended use is to have a different variable values for deployment.

The first use case is for the 'url' config variable. Creating an
additional _config-deploy.yml file with the deployment URL, one can
transparently use development and production URLs.