Add addtional CLI option --branch-from
This supports a little more flexibility in which
branches can be used for deploying and building.

Specifically, it lets a user build from some non-master
branch and push to a master branch, which is useful
if you're deploying to the standard myname.github.io.

Example:

jgd --branch master --branch-from source